changes file structure. Moves files into folders.
