--- src/main/java/com/myhomebe/ProjectRepository.java
@@ -1,8 +1,0 @@
-package com.myhomebe;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-interface ProjectRepository extends JpaRepository<Project, Long> {
-}
