+++ src/main/java/com/myhomebe/controller/HomeController.java
@@ -1,4 +1,4 @@
-package com.myhomebe;
+package com.myhomebe.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.web.bind.annotation.RequestMapping;
