+++ src/main/java/com/myhomebe/repository/ProjectRepository.java
@@ -1,0 +1,9 @@
+package com.myhomebe.repository;
+
+import com.myhomebe.model.Project;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProjectRepository extends JpaRepository<Project, Long> {
+}
